Adds Forum resource
Added a route to forums,
added a model for forum,
added a controller for forum,
forum belongs to User with 'populate',
tested user and forum resources with curl scripts.

All CRUD works except destroy because I don't want users to be able
to destroy forums, only comments.

FUTURE: do all this for 'comments'